Changes regarding renaming of scalismo-hdf-json to scalismo-hdf5-json
