+++ build.sbt
@@ -64,7 +64,7 @@ lazy val root = (project in file("."))
       ("org.scalanlp" %% "breeze-natives" % "2.1.0"),
       "ch.unibas.cs.gravis" % "scalismo-niftijiojar" % "0.1.0",
       //"com.lihaoyi" %% "upickle" % "2.0.0",
-      ("ch.unibas.cs.gravis" %% "scalismo-hdf-json" % "0.1.0-SNAPSHOT").cross(CrossVersion.for2_13Use3),
+      ("ch.unibas.cs.gravis" %% "scalismo-hdf5-json" % "0.1.0-SNAPSHOT").cross(CrossVersion.for2_13Use3),
       "ch.unibas.cs.gravis" % "vtkjavanativesall" % "0.1.1",
       "io.jhdf" % "jhdf" % "0.6.9",
       "org.slf4j" % "slf4j-nop" % "1.6.0" // this silences slf4j complaints in registration classes
