+++ src/main/scala/scalismo/io/statisticalmodel/HDF5JsonReader.scala
@@ -16,13 +16,13 @@
 package scalismo.io.statisticalmodel
 
 import io.jhdf.HdfFile
-import scalismo.hdfjson.{FloatArray1D, FloatArray2D, HDFJson, HDFPath, IntArray1D, IntArray2D}
+import scalismo.hdf5json.{FloatArray1D, FloatArray2D, HDF5Json, HDFPath, IntArray1D, IntArray2D}
 
 import java.io.Closeable
 import scala.jdk.CollectionConverters.MapHasAsScala
 import scala.util.{Failure, Success, Try}
 
-class HDF5JsonReader(hdfjson: HDFJson) extends Closeable with StatisticalModelReader {
+class HDF5JsonReader(hdfjson: HDF5Json) extends Closeable with StatisticalModelReader {
 
   override def close(): Unit = {
     /* there is no need to close this file, as hdfjson is  string */
